fix: handle symbolinfo
closes: #8

An LSP can return either a SymbolInformation structure or a
DocumentSymbol structure when a documentSymbol request is made.

The new method is exported in 'litee.lib.lsp' and requires a pull of the
latest lib.

This commit detects if a SymbolInformation structure is returned and
converts it to a DocumentSymbol so the rest of the code need not change.