useability: update `invoking_win` on symboltree refresh
prior to this commit the `invoking_win` field in symboltree's state was
never updated when refreshing a symboltree.

this causes jumps and auto-highlights from the symboltree to work on
stale windows.

after this commit, browsing and jumping via the symboltree will always
effect the window in which the symboltree refers to.

Signed-off-by: ldelossa <louis.delos@gmail.com>